services/normalization_service/requirements.txt
kafka-python
langchain-groq
